implemented most of session creation and agreemnt, remains some polishing,and remains to add payment servie and connect it with banking api

+++ learn-bridge-back-end/src/main/java/com/learnbridge/learn_bridge_back_end/dao/NotificationsDAOImpl.java
@@ -1,6 +1,9 @@
 package com.learnbridge.learn_bridge_back_end.dao;
 
+import com.learnbridge.learn_bridge_back_end.entity.Agreement;
 import com.learnbridge.learn_bridge_back_end.entity.Notifications;
+import com.learnbridge.learn_bridge_back_end.entity.ReadStatus;
+import com.learnbridge.learn_bridge_back_end.entity.User;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.TypedQuery;
@@ -17,47 +20,72 @@ public class NotificationsDAOImpl implements NotificationsDAO {
 
     @Override
     @Transactional
-    public void saveNotification(Notifications notification) {
+    public Notifications saveNotification(Notifications notification) {
         entityManager.persist(notification);
+        return notification;
     }
 
     @Override
     @Transactional
-    public void updateNotification(Notifications notification) {
-
-        entityManager.merge(notification);
+    public Notifications updateNotification(Notifications notification) {
+        return entityManager.merge(notification);
     }
 
     @Override
     @Transactional
     public void deleteNotification(Long notificationId) {
         Notifications notification = entityManager.find(Notifications.class, notificationId);
-        entityManager.remove(entityManager.contains(notification) ? notification : entityManager.merge(notification));
+        if (notification != null) {
+            entityManager.remove(entityManager.contains(notification) ? notification : entityManager.merge(notification));
+        }
     }
 
     @Override
     public List<Notifications> findAllNotifications() {
-
         return entityManager.createQuery("from Notifications", Notifications.class).getResultList();
     }
 
     @Override
     @Transactional
     public void deleteNotificationsByUserId(Long userId) {
-
-        List<Notifications> notificationsToBeDeleted = findNotificationByUserId(userId);
-
+        List<Notifications> notificationsToBeDeleted = findNotificationsByUserId(userId);
         for (Notifications notification : notificationsToBeDeleted) {
             deleteNotification(notification.getNotificationId());
         }
     }
 
-    @Override
-    public List<Notifications> findNotificationByUserId(Long userId) {
-
+    public List<Notifications> findNotificationsByUserId(Long userId) {
         String sqlStatement = "from Notifications n where n.user.userId = :userId";
         TypedQuery<Notifications> query = entityManager.createQuery(sqlStatement, Notifications.class);
         query.setParameter("userId", userId);
+        return query.getResultList();
+    }
+
+    @Override
+    public List<Notifications> findNotificationsByUserId(User user) {
+        TypedQuery<Notifications> query = entityManager.createQuery(
+                "SELECT n FROM Notifications n WHERE n.user = :user ORDER BY n.timestamp DESC",
+                Notifications.class);
+        query.setParameter("user", user);
+        return query.getResultList();
+    }
+
+    @Override
+    public List<Notifications> findNotificationsByAgreement(Agreement agreement) {
+        TypedQuery<Notifications> query = entityManager.createQuery(
+                "SELECT n FROM Notifications n WHERE n.agreement = :agreement",
+                Notifications.class);
+        query.setParameter("agreement", agreement);
+        return query.getResultList();
+    }
+
+    @Override
+    public List<Notifications> findUnreadNotificationsByUser(User user) {
+        TypedQuery<Notifications> query = entityManager.createQuery(
+                "SELECT n FROM Notifications n WHERE n.user = :user AND n.readStatus = :readStatus ORDER BY n.timestamp DESC",
+                Notifications.class);
+        query.setParameter("user", user);
+        query.setParameter("readStatus", ReadStatus.UNREAD);
         return query.getResultList();
     }
